Move processor from __init__ and add some data augmentation things
